removed animation
decided to keep it plain and simple; i'm going to use links and redirect users to detail pages instead of using JS to change the content of the homepage.

this way I don't have to write a lot of redundant code to make it work and it has the benefit of refreshing the page.